backend_django/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
google-generativeai
pypdf
